+++ sktime/utils/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities for adapters and dispatchers."""
